+++ core/exchange/src/main/kotlin/com/example/currencyconverter/core/exchange/converter/model/ExchangeTransaction.kt
@@ -2,10 +2,11 @@ package com.example.currencyconverter.core.exchange.converter.model
 
 import com.example.currencyconverter.core.exchange.rates.model.Currency
 import com.example.currencyconverter.core.exchange.rates.model.ExchangeRate
+import java.math.BigDecimal
 
 data class ExchangeTransaction(
     val base: Currency,
-    val baseBalance: Double,
-    val amount: Double,
+    val baseBalance: BigDecimal,
+    val amount: BigDecimal,
     val rate: ExchangeRate,
 )
