+++ core/exchange/src/main/kotlin/com/example/currencyconverter/core/exchange/fees/FeeWeight.kt
@@ -1,7 +1,9 @@
 package com.example.currencyconverter.core.exchange.fees
 
+import java.math.BigDecimal
+
 internal sealed interface FeeWeight {
-    data class SubstituteAll(val amount: Double) : FeeWeight
-    data class Accumulative(val amount: Double) : FeeWeight
+    data class SubstituteAll(val amount: BigDecimal) : FeeWeight
+    data class Accumulative(val amount: BigDecimal) : FeeWeight
     data object NotValid : FeeWeight
 }
