+++ features/exchange/src/main/kotlin/com/example/currencyconverter/features/exchange/ui/NumberFormatter.kt
@@ -1,11 +1,12 @@
 package com.example.currencyconverter.features.exchange.ui
 
+import java.math.BigDecimal
 import java.util.Locale
 
 internal object NumberFormatter {
 
-    fun format(value: Double): String {
-        val fractionDigitCount = if ((value * 100).toInt() > 0) {
+    fun format(value: BigDecimal): String {
+        val fractionDigitCount = if ((value * BigDecimal(100.0)).toInt() > 0) {
             2
         } else {
             6
